requirements.txt
matplotlib
numpy
Pillow
pycocotools
tqdm
transformers
opencv-python
scipy